backend/requirements.txt
fastapi
uvicorn
pandas
matplotlib
langchain
langchain-experimental
langchain-google-genai
python-dotenv
tabulate
tempfile